switchable view (table/chart) for data page
