+++ common/types.ts
@@ -270,6 +270,10 @@ export interface DatapointsByMetric {
   [metricKey: string]: DatapointsGroupedByAggregateAndDisaggregations;
 }
 
+export interface RawDatapointsByMetric {
+  [metricKey: string]: RawDatapoint[];
+}
+
 export type DataVizTimeRange = 0 | 6 | 12 | 60 | 120;
 
 export const DataVizTimeRangesMap: { [key: string]: DataVizTimeRange } = {
