+++ publisher/src/components/ReviewMetrics/ReviewMetrics.styles.tsx
@@ -111,109 +111,3 @@ export const SectionTitleMonths = styled.div`
 export const SectionTitleOverwrites = styled(SectionTitleMonths)`
   background-color: ${palette.solid.orange};
 `;
-
-export const DatapointsTableContainer = styled.div`
-  margin-top: 16px;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-`;
-
-export const DatapointsTableNamesContainer = styled.div`
-  width: 240px;
-  min-width: 240px;
-  padding-top: 33px;
-`;
-
-export const DatapointsTableNamesTable = styled.table`
-  border-collapse: collapse;
-`;
-
-export const DatapointsTableNamesRow = styled.tr``;
-
-export const DatapointsTableNamesCell = styled.td`
-  padding-top: 4px;
-  padding-bottom: 4px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  font-size: 18px;
-  font-weight: 400;
-  max-width: 232px;
-  height: 32px;
-`;
-export const DatapointsTableNamesDivider = styled.td`
-  ${typography.sizeCSS.small}
-  padding-top: 8px;
-  padding-bottom: 8px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
-
-export const DatapointsTableDetailsContainer = styled.div`
-  position: relative;
-  display: flex;
-  max-width: 624px;
-`;
-export const DatapointsTableDetailScrollContainer = styled.div`
-  overflow-x: scroll;
-  padding-bottom: 10px;
-`;
-export const DatapointsTableDetailsContainerOverlay = styled.div`
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  pointer-events: none;
-`;
-export const DatapointsTableDetailsContainerOverlayLeftGradient = styled.div`
-  background: linear-gradient(
-    0.25turn,
-    rgb(255, 255, 255, 1),
-    rgb(255, 255, 255, 0)
-  );
-  width: 32px;
-  height: 100%;
-`;
-export const DatapointsTableDetailsContainerOverlayRightGradient = styled.div`
-  background: linear-gradient(
-    0.25turn,
-    rgb(255, 255, 255, 0),
-    rgb(255, 255, 255, 1)
-  );
-  width: 32px;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  right: 0;
-`;
-export const DatapointsTableDetailsTable = styled.table`
-  border-collapse: collapse;
-  margin: 0 32px;
-`;
-export const DatapointsTableDetailsRowHead = styled.thead``;
-export const DatapointsTableDetailsRowBody = styled.tbody``;
-export const DatapointsTableDetailsRow = styled.tr``;
-export const DatapointsTableDetailsRowHeader = styled.th`
-  ${typography.sizeCSS.small}
-  padding-bottom: 8px;
-  padding-top: 8px;
-  text-align: center;
-  border: 1px solid ${palette.highlight.grey3};
-`;
-export const DatapointsTableDetailsCell = styled.td`
-  padding-left: 32px;
-  padding-right: 32px;
-  padding-bottom: 4px;
-  padding-top: 4px;
-  font-size: 18px;
-  height: 32px;
-  font-weight: 400;
-  text-align: center;
-  border: 1px solid ${palette.highlight.grey3};
-  white-space: nowrap;
-`;
-export const DatapointsTableDetailsDivider = styled.tr`
-  height: 32px;
-  border: 1px solid ${palette.highlight.grey3};
-`;
