+++ publisher/src/components/ReviewMetrics/ReviewMetrics.tsx
@@ -15,44 +15,17 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import {
-  DataVizAggregateName,
-  RawDatapoint,
-} from "@justice-counts/common/types";
+import { DatapointsTableView } from "@justice-counts/common/components/DataViz/DatapointsTableView";
 import { observer } from "mobx-react-lite";
 import React, { useEffect } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
 import logoImg from "../assets/jc-logo-vector.png";
-import {
-  Button,
-  DataUploadHeader,
-  OrangeText,
-  StrikethroughText,
-} from "../DataUpload/DataUpload.styles";
+import { Button, DataUploadHeader } from "../DataUpload";
 import { UploadedMetric } from "../DataUpload/types";
-import { formatDateShort, sortDatapointDimensions } from "../DataViz/utils";
 import { Logo, LogoContainer } from "../Header";
 import {
   Container,
-  DatapointsTableContainer,
-  DatapointsTableDetailsCell,
-  DatapointsTableDetailsContainer,
-  DatapointsTableDetailsContainerOverlay,
-  DatapointsTableDetailsContainerOverlayLeftGradient,
-  DatapointsTableDetailsContainerOverlayRightGradient,
-  DatapointsTableDetailScrollContainer,
-  DatapointsTableDetailsDivider,
-  DatapointsTableDetailsRow,
-  DatapointsTableDetailsRowBody,
-  DatapointsTableDetailsRowHead,
-  DatapointsTableDetailsRowHeader,
-  DatapointsTableDetailsTable,
-  DatapointsTableNamesCell,
-  DatapointsTableNamesContainer,
-  DatapointsTableNamesDivider,
-  DatapointsTableNamesRow,
-  DatapointsTableNamesTable,
   Heading,
   MainPanel,
   SectionContainer,
@@ -63,14 +36,6 @@ import {
   SectionTitleOverwrites,
   Subheading,
 } from "./ReviewMetrics.styles";
-
-type AggregationRowData = (RawDatapoint | undefined)[];
-
-type DisaggregationRowData = {
-  [disaggregation: string]: {
-    [dimension: string]: (RawDatapoint | undefined)[];
-  };
-};
 
 const ReviewMetrics: React.FC = observer(() => {
   const location = useLocation();
@@ -102,52 +67,9 @@ const ReviewMetrics: React.FC = observer(() => {
     // keep track of count of overwritten values
     let overwrittenValuesCount = 0;
 
-    // Create array of aggregate values, each value indexed with their corresponding date column
-    const aggregateRowData: AggregationRowData = new Array(
-      startDates.length
-    ).fill(undefined);
-
-    // Create map of disaggregations and dimensions, each dimension containing array of values,
-    // each value indexed with their corresponding date column
-    const disaggregationRowData: DisaggregationRowData = {};
-
-    // create a mapping from start date to the column index the start date is located in for fast lookup
-    const startDatesIndexLookup = startDates.reduce((map, current, idx) => {
-      map[current] = idx; /* eslint-disable-line no-param-reassign */
-      return map;
-    }, {} as { [key: string]: number });
-
     metric.datapoints.forEach((dp) => {
-      if (dp.disaggregation_display_name && dp.dimension_display_name) {
-        if (!disaggregationRowData[dp.disaggregation_display_name]) {
-          disaggregationRowData[dp.disaggregation_display_name] = {};
-        }
-        if (
-          !disaggregationRowData[dp.disaggregation_display_name][
-            dp.dimension_display_name
-          ]
-        ) {
-          disaggregationRowData[dp.disaggregation_display_name][
-            dp.dimension_display_name
-          ] = new Array(startDates.length).fill(undefined);
-        }
-        if (
-          !disaggregationRowData[dp.disaggregation_display_name][
-            dp.dimension_display_name
-          ][startDatesIndexLookup[dp.start_date]]
-        ) {
-          disaggregationRowData[dp.disaggregation_display_name][
-            dp.dimension_display_name
-          ][startDatesIndexLookup[dp.start_date]] = dp;
-          if (dp.old_value !== null) {
-            overwrittenValuesCount += 1;
-          }
-        }
-      } else if (!aggregateRowData[startDatesIndexLookup[dp.start_date]]) {
-        aggregateRowData[startDatesIndexLookup[dp.start_date]] = dp;
-        if (dp.old_value !== null) {
-          overwrittenValuesCount += 1;
-        }
+      if (dp.old_value !== null) {
+        overwrittenValuesCount += 1;
       }
     });
 
@@ -168,143 +90,8 @@ const ReviewMetrics: React.FC = observer(() => {
             {startDates.length !== 1 ? "s" : ""}
           </SectionTitleMonths>
         </SectionTitleContainer>
-        {renderDatapointsTable(
-          aggregateRowData,
-          disaggregationRowData,
-          startDates
-        )}
+        <DatapointsTableView datapoints={metric.datapoints} />
       </SectionContainer>
-    );
-  };
-
-  const renderDatapointsTable = (
-    aggregateRowData: AggregationRowData,
-    disaggregationRowData: DisaggregationRowData,
-    startDates: string[]
-  ) => {
-    /**
-     * The Datapoints Table is made up of a few parts:
-     * - Column of disaggregations and dimensions that remains in place as
-     *   the rest of the table scrolls horizontally
-     * - HTML table containing:
-     *   - a table header row of report start dates, each date represents a column title
-     *   - a table row of the aggregate values, each value corresponds with the date column it is reported in
-     *   - empty table rows used to space apart different disaggregations
-     *   - table rows for each dimension, each value corresponds with the date column it is reported in
-     */
-
-    return (
-      <DatapointsTableContainer>
-        <DatapointsTableNamesContainer>
-          <DatapointsTableNamesTable>
-            <DatapointsTableNamesRow>
-              <DatapointsTableNamesCell>
-                {DataVizAggregateName}
-              </DatapointsTableNamesCell>
-            </DatapointsTableNamesRow>
-            {Object.entries(disaggregationRowData).map(
-              ([disaggregation, dimension]) => (
-                <React.Fragment key={disaggregation}>
-                  <DatapointsTableNamesRow>
-                    <DatapointsTableNamesDivider>
-                      {disaggregation}
-                    </DatapointsTableNamesDivider>
-                  </DatapointsTableNamesRow>
-                  {Object.keys(dimension)
-                    .sort(sortDatapointDimensions)
-                    .map((dimensionName) => (
-                      <DatapointsTableNamesRow key={dimensionName}>
-                        <DatapointsTableNamesCell>
-                          {dimensionName}
-                        </DatapointsTableNamesCell>
-                      </DatapointsTableNamesRow>
-                    ))}
-                </React.Fragment>
-              )
-            )}
-          </DatapointsTableNamesTable>
-        </DatapointsTableNamesContainer>
-        <DatapointsTableDetailsContainer>
-          <DatapointsTableDetailScrollContainer>
-            <DatapointsTableDetailsTable>
-              <DatapointsTableDetailsRowHead>
-                <DatapointsTableDetailsRow>
-                  {startDates.map((date) => (
-                    <DatapointsTableDetailsRowHeader key={date}>
-                      {formatDateShort(date)}
-                    </DatapointsTableDetailsRowHeader>
-                  ))}
-                </DatapointsTableDetailsRow>
-              </DatapointsTableDetailsRowHead>
-              <DatapointsTableDetailsRowBody>
-                <DatapointsTableDetailsRow>
-                  {aggregateRowData.map((dp, index) =>
-                    // row data could be null, so no distinct key given in that case
-                    dp === undefined
-                      ? renderDatapointsValue(index, null, null)
-                      : renderDatapointsValue(index, dp.value, dp.old_value)
-                  )}
-                </DatapointsTableDetailsRow>
-                {Object.entries(disaggregationRowData).map(
-                  ([disaggregation, dimension]) => (
-                    <React.Fragment key={disaggregation}>
-                      <DatapointsTableDetailsDivider />
-                      {Object.entries(dimension)
-                        .sort(([a], [b]) => sortDatapointDimensions(a, b))
-                        .map(([key, dps]) => (
-                          <DatapointsTableDetailsRow key={key}>
-                            {dps.map((dp, index) =>
-                              dp === undefined
-                                ? renderDatapointsValue(index, null, null)
-                                : renderDatapointsValue(
-                                    index,
-                                    dp.value,
-                                    dp.old_value
-                                  )
-                            )}
-                          </DatapointsTableDetailsRow>
-                        ))}
-                    </React.Fragment>
-                  )
-                )}
-              </DatapointsTableDetailsRowBody>
-            </DatapointsTableDetailsTable>
-          </DatapointsTableDetailScrollContainer>
-          <DatapointsTableDetailsContainerOverlay>
-            <DatapointsTableDetailsContainerOverlayLeftGradient />
-            <DatapointsTableDetailsContainerOverlayRightGradient />
-          </DatapointsTableDetailsContainerOverlay>
-        </DatapointsTableDetailsContainer>
-      </DatapointsTableContainer>
-    );
-  };
-
-  const renderDatapointsValue = (
-    key: React.Key,
-    value: string | number | null,
-    oldValue: string | number | null
-  ) => {
-    // Uncomment this for special treatment when a datapoint is newly added
-    // as opposed to returning the same value
-    // if (oldValue === null && value !== null) {
-    //   return (
-    //     <DatapointsTableDetailsCell key={key}>
-    //       {value} NEW
-    //     </DatapointsTableDetailsCell>
-    //   );
-    // }
-    if (oldValue !== null && oldValue !== value) {
-      return (
-        <DatapointsTableDetailsCell key={key}>
-          <StrikethroughText>{oldValue}</StrikethroughText>
-          {` ➞ `}
-          {value}
-          <OrangeText>*</OrangeText>
-        </DatapointsTableDetailsCell>
-      );
-    }
-    return (
-      <DatapointsTableDetailsCell key={key}>{value}</DatapointsTableDetailsCell>
     );
   };
